// src/generated/serverEnv.d.ts
export declare const SERVER_ENV: {
  AUTH_SECRET: string | undefined;
  NEXTAUTH_SECRET: string | undefined;
  AUTH_URL: string | undefined;
  NEXTAUTH_URL: string | undefined;
  AUTH_TRUST_HOST: string | undefined;
  GOOGLE_CLIENT_ID: string | undefined;
  AUTH_GOOGLE_ID: string | undefined;
  GOOGLE_CLIENT_SECRET: string | undefined;
  AUTH_GOOGLE_SECRET: string | undefined;
  OVERRIDE_EMAIL: string | undefined;
  GOOGLE_SERVICE_ACCOUNT_TYPE: string | undefined;
  GOOGLE_SERVICE_ACCOUNT_PROJECT_ID: string | undefined;
  GOOGLE_SERVICE_ACCOUNT_PRIVATE_KEY_ID: string | undefined;
  GOOGLE_SERVICE_ACCOUNT_PRIVATE_KEY: string | undefined;
  GOOGLE_SERVICE_ACCOUNT_CLIENT_EMAIL: string | undefined;
  GOOGLE_SERVICE_ACCOUNT_CLIENT_ID: string | undefined;
  GOOGLE_DRIVE_CLIENT_ID: string | undefined;
  GOOGLE_DRIVE_CLIENT_SECRET: string | undefined;
  GOOGLE_DRIVE_REFRESH_TOKEN: string | undefined;
  GOOGLE_DRIVE_REDIRECT_URI: string | undefined;
  R2_ACCOUNT_ID: string | undefined;
  R2_ACCESS_KEY_ID: string | undefined;
  R2_SECRET_ACCESS_KEY: string | undefined;
  R2_BUCKET: string | undefined;
  R2_PUBLIC_BASE_URL: string | undefined;
  SHEET_NAME_ENTRADAS: string | undefined;
  NEXT_PUBLIC_SPREADSHEET_ID: string | undefined;
};